split AuthDomainSpec into 2 tests for faster time
